moved logout to correct location, deleted money textview
