+++ app/src/main/java/com/es/findsoccerplayers/LocationAccessActivity.java
@@ -1,23 +1,15 @@
 package com.es.findsoccerplayers;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
 
 import android.Manifest;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
-import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.net.Uri;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
-import android.widget.Toast;
-
-import com.es.findsoccerplayers.position.MapElements;
-import com.es.findsoccerplayers.position.PositionClient;
 
 public class LocationAccessActivity extends AppCompatActivity {
 
@@ -27,7 +19,7 @@ public class LocationAccessActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_location_access);
+        setContentView(R.layout.act_location_access);
 
         next = findViewById(R.id.next_text);
